# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "summary": "Connexion utilisateur",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès"
          },
          "400": {
            "description": "Données invalides ou email déjà utilisé"
          }
        },
        "summary": "Enregistrement d'un nouvel utilisateur",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/attachments/upload": {
      "post": {
        "operationId": "AttachmentsController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "deadlineId": {
                    "type": "string"
                  },
                  "classification": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pièce jointe créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "404": {
            "description": "Échéance non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Uploader une nouvelle pièce jointe",
        "tags": [
          "attachments"
        ]
      }
    },
    "/attachments": {
      "get": {
        "operationId": "AttachmentsController_findAll",
        "parameters": [
          {
            "name": "deadlineId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploaderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classification",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des pièces jointes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer toutes les pièces jointes",
        "tags": [
          "attachments"
        ]
      }
    },
    "/attachments/{id}": {
      "get": {
        "operationId": "AttachmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la pièce jointe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pièce jointe trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "404": {
            "description": "Pièce jointe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une pièce jointe par son ID",
        "tags": [
          "attachments"
        ]
      },
      "patch": {
        "operationId": "AttachmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la pièce jointe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttachmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pièce jointe mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé"
          },
          "404": {
            "description": "Pièce jointe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une pièce jointe",
        "tags": [
          "attachments"
        ]
      },
      "delete": {
        "operationId": "AttachmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la pièce jointe",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pièce jointe supprimée"
          },
          "403": {
            "description": "Accès non autorisé"
          },
          "404": {
            "description": "Pièce jointe non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une pièce jointe",
        "tags": [
          "attachments"
        ]
      }
    },
    "/attachments/deadline/{deadlineId}": {
      "get": {
        "operationId": "AttachmentsController_findByDeadline",
        "parameters": [
          {
            "name": "deadlineId",
            "required": true,
            "in": "path",
            "description": "ID de l'échéance",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des pièces jointes de l'échéance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les pièces jointes par échéance",
        "tags": [
          "attachments"
        ]
      }
    },
    "/attachments/uploader/{uploaderId}": {
      "get": {
        "operationId": "AttachmentsController_findByUploader",
        "parameters": [
          {
            "name": "uploaderId",
            "required": true,
            "in": "path",
            "description": "ID de l'uploader",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des pièces jointes de l'uploader",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Attachment"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer les pièces jointes par uploader",
        "tags": [
          "attachments"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestion d'Échéances",
    "description": "API pour l'application de gestion d'échéances du secteur défense",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {}
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterDto": {
        "type": "object",
        "properties": {}
      },
...
      "Attachment": {
        "type": "object",
        "properties": {}
      },
      "UpdateAttachmentDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Nom du fichier",
            "example": "rapport_mensuel.pdf"
          },
          "mimeType": {
            "type": "string",
            "description": "Type MIME du fichier",
            "example": "application/pdf"
          },
          "size": {
            "type": "number",
            "description": "Taille du fichier en octets",
            "example": 1048576
          },
          "path": {
            "type": "string",
            "description": "Chemin de stockage du fichier",
            "example": "uploads/2025/04/rapport_mensuel_123456.pdf"
          },
          "deadlineId": {
            "type": "string",
            "description": "ID de l'échéance associée",
            "example": "7a9d0f35-d562-4219-c732-d35cf1a2e5f9"
          },
          "uploaderId": {
            "type": "string",
            "description": "ID de l'utilisateur ayant ajouté la pièce jointe",
            "example": "5f8d0f55-e477-4915-b522-c21ba1f70d9e"
          },
          "classification": {
            "type": "string",
            "description": "Niveau de classification du document",
            "example": "Confidentiel"
          }
        }
      },
...
    }
  }
}